feat(ui): add Customize screen for Tools and MCP, clean up Settings
- New /customize route with Puzzle icon in the nav rail
- CustomizeScreen: Web Search (Tavily key) + MCP Servers (add/edit/delete/toggle)
- Removed ToolsSection and McpSection from SettingsScreen — no longer clutters settings
- Settings remains focused on engine, model, network, startup, and persona config
